Create extract_signals module
